Add files via upload
My initial thoughts for the world I want to build. More will be added to this file and there will be other files pertaining to the rest of the world. These other files will be organized based off of the type of information they have, example would be government, general population, notable families and figure and etc....